Use StbImageSharp for WindowsDX, UWP and Android(#6867)
Replaced PngWriter and native texture saving/loading so there's only one codepath to maintain.
