+++ MonoGame.Framework/Platform/Graphics/Texture2D.OpenGL.cs
@@ -6,8 +6,6 @@ using System;
 using System.IO;
 using System.Runtime.InteropServices;
 using MonoGame.Utilities;
-using MonoGame.Utilities.Png;
-using StbImageSharp;
 
 #if IOS
 using UIKit;
@@ -20,7 +18,6 @@ using System.Drawing;
 using MonoGame.OpenGL;
 using GLPixelFormat = MonoGame.OpenGL.PixelFormat;
 using PixelFormat = MonoGame.OpenGL.PixelFormat;
-using StbImageWriteSharp;
 
 #if ANDROID
 using Android.Graphics;
@@ -257,48 +254,6 @@ namespace Microsoft.Xna.Framework.Graphics
 #endif
         }
 
-        private unsafe static Texture2D PlatformFromStream(GraphicsDevice graphicsDevice, Stream stream)
-        {
-            byte[] bytes;
-
-            // Rewind stream if it is at end
-            if (stream.CanSeek && stream.Length == stream.Position)
-            {
-                stream.Seek(0, SeekOrigin.Begin);
-            }
-
-            // Copy it's data to memory
-            // As some platforms dont provide full stream functionality and thus streams can't be read as it is
-            using (var ms = new MemoryStream())
-            {
-                stream.CopyTo(ms);
-                bytes = ms.ToArray();
-            }
-
-            // The data returned is always four channel BGRA
-            var result = ImageResult.FromMemory(bytes, StbImageSharp.ColorComponents.RedGreenBlueAlpha);
-
-            // XNA blacks out any pixels with an alpha of zero.
-            fixed (byte* b = &result.Data[0])
-            {
-                for (var i = 0; i < result.Data.Length; i += 4)
-                {
-                    if (b[i + 3] == 0)
-                    {
-                        b[i + 0] = 0;
-                        b[i + 1] = 0;
-                        b[i + 2] = 0;
-                    }
-                }
-            }
-
-            Texture2D texture = null;
-            texture = new Texture2D(graphicsDevice, result.Width, result.Height);
-            texture.SetData(result.Data);
-
-            return texture;
-        }
-
 #if IOS
         [CLSCompliant(false)]
         public static Texture2D FromStream(GraphicsDevice graphicsDevice, UIImage uiImage)
@@ -428,100 +383,6 @@ namespace Microsoft.Xna.Framework.Graphics
             }
 #endif
         }
-
-#if DESKTOPGL
-        internal enum ImageWriterFormat
-        {
-            Jpg,
-            Png
-        }
-#endif
-
-        private void PlatformSaveAsJpeg(Stream stream, int width, int height)
-        {
-#if DESKTOPGL
-            SaveAsImage(stream, width, height, ImageWriterFormat.Jpg);
-#elif ANDROID
-            SaveAsImage(stream, width, height, Bitmap.CompressFormat.Jpeg);
-#else
-            throw new NotImplementedException();
-#endif
-        }
-
-        private void PlatformSaveAsPng(Stream stream, int width, int height)
-        {
-#if DESKTOPGL
-            SaveAsImage(stream, width, height, ImageWriterFormat.Png);
-#elif ANDROID
-            SaveAsImage(stream, width, height, Bitmap.CompressFormat.Png);
-#else
-            var pngWriter = new PngWriter();
-            pngWriter.Write(this, stream);
-#endif
-        }
-
-#if DESKTOPGL
-        internal unsafe void SaveAsImage(Stream stream, int width, int height, ImageWriterFormat format)
-        {
-	        if (stream == null)
-	        {
-		        throw new ArgumentNullException("stream", "'stream' cannot be null (Nothing in Visual Basic)");
-	        }
-	        if (width <= 0)
-	        {
-		        throw new ArgumentOutOfRangeException("width", width, "'width' cannot be less than or equal to zero");
-	        }
-	        if (height <= 0)
-	        {
-		        throw new ArgumentOutOfRangeException("height", height, "'height' cannot be less than or equal to zero");
-	        }
-	        Color[] data = null;
-	        try
-	        {
-                data = GetColorData();
-
-                // Write
-                fixed (Color* ptr = &data[0])
-                {
-                    var writer = new ImageWriter();
-                    switch (format)
-                    {
-                        case ImageWriterFormat.Jpg:
-                            writer.WriteJpg(ptr, width, height, StbImageWriteSharp.ColorComponents.RedGreenBlueAlpha, stream, 90);
-                            break;
-                        case ImageWriterFormat.Png:
-                            writer.WritePng(ptr, width, height, StbImageWriteSharp.ColorComponents.RedGreenBlueAlpha, stream);
-                            break;
-                    }
-                }
-            }
-            finally
-	        {
-		        if (data != null)
-		        {
-			        data = null;
-		        }
-	        }
-        }
-#elif ANDROID
-        private void SaveAsImage(Stream stream, int width, int height, Bitmap.CompressFormat format)
-        {
-            int[] data = new int[width * height];
-            GetData(data);
-            // internal structure is BGR while bitmap expects RGB
-            for (int i = 0; i < data.Length; ++i)
-            {
-                uint pixel = (uint)data[i];
-                data[i] = (int)((pixel & 0xFF00FF00) | ((pixel & 0x00FF0000) >> 16) | ((pixel & 0x000000FF) << 16));
-            }
-            using (Bitmap bitmap = Bitmap.CreateBitmap(width, height, Bitmap.Config.Argb8888))
-            {
-                bitmap.SetPixels(data, 0, width, 0, 0, width, height);
-                bitmap.Compress(format, 100, stream);
-                bitmap.Recycle();
-            }
-        }
-#endif
 
         // This method allows games that use Texture2D.FromStream
         // to reload their textures after the GL context is lost.
